+++ safe/app/src/main/java/com/abb/safe/Fragment/sRoutesFragment.java
@@ -9,7 +9,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.abb.safe.MapsActivity;
 import com.abb.safe.R;
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -23,10 +22,8 @@ import com.google.android.gms.maps.model.PolylineOptions;
 import com.google.android.gms.maps.model.RoundCap;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-import okhttp3.Route;
 
 
 public class sRoutesFragment extends Fragment implements OnMapReadyCallback {
