+++ safe/app/src/main/java/com/abb/safe/SettingActivity.java
@@ -117,6 +117,7 @@ public class SettingActivity extends AppCompatActivity {
             public void onClick(View view) {
                 Intent intent = new Intent(getApplicationContext(),MapsActivity.class);
                 startActivity(intent);
+                finish();
             }
         });
         screenroute.setOnClickListener(new View.OnClickListener() {
@@ -124,6 +125,7 @@ public class SettingActivity extends AppCompatActivity {
             public void onClick(View view) {
                 Intent intent = new Intent(getApplicationContext(),RoutesActivity.class);
                 startActivity(intent);
+                finish();
             }
         });
         screensetting.setOnClickListener(new View.OnClickListener() {
@@ -131,6 +133,7 @@ public class SettingActivity extends AppCompatActivity {
             public void onClick(View view) {
                 Intent intent = new Intent(getApplicationContext(),SettingActivity.class);
                 startActivity(intent);
+                finish();
             }
         });
